feat: move GlobalKeystoreDialog to @swapkit/ui/react, show when connecting keystore
