+++ playgrounds/nextjs/src/components/providers.tsx
@@ -2,7 +2,6 @@
 
 import type { PropsWithChildren } from "react";
 import { ThemeProvider } from "./containers/Theme";
-import { GlobalKeystoreDialog } from "./GlobalKeystoreDialog";
 import { Toaster } from "./ui/sonner";
 import { TooltipProvider } from "./ui/tooltip";
 
@@ -12,7 +11,6 @@ export function AppProviders({ children }: PropsWithChildren) {
       <TooltipProvider>
         <div className="mx-auto max-w-1/2">{children}</div>
         <Toaster position="bottom-right" />
-        <GlobalKeystoreDialog />
       </TooltipProvider>
     </ThemeProvider>
   );
